+++ ios/Classes/DoozMeshManagerApi/Channels/Arguments/DoozEpochSetArguments.swift
@@ -1,0 +1,18 @@
+//
+//  DoozEpochSetArguments.swift
+//  nordic_nrf_mesh
+//
+//  Created by OZEO DOOZ on 11/02/2022.
+//
+
+import Foundation
+
+struct DoozEpochSetArguments: BaseFlutterArguments {
+    let tzData: UInt16
+    let command: UInt8
+    let io: UInt8
+    let unused: UInt8
+    let epoch: UInt32
+    let correlation: UInt32
+    let extra: UInt8?
+}