+++ ios/Classes/ModelDelegate/GenericLevelServerDelegate.swift
@@ -75,7 +75,8 @@ class GenericLevelServerDelegate: ModelDelegate {
             GenericMoveSet.self,
             GenericMoveSetUnacknowledged.self,
             MagicLevelGet.self,
-            MagicLevelSet.self
+            MagicLevelSet.self,
+            DoozEpochStatus.self
         ]
         messageTypes = types.toMap()
     }
@@ -163,7 +164,17 @@ class GenericLevelServerDelegate: ModelDelegate {
             lastTransaction = (source: source, destination: destination, tid: request.tid, timestamp: Date())
             
             return MagicLevelSetStatus(io: request.mIO, index: request.mIndex, value: request.mValue, correlation: request.mCorrelation, tId: request.tid)
-            
+        case let request as DoozEpochSet:
+            // Ignore a repeated request (with the same TID) from the same source
+            // and sent to the same destinatino when it was received within 6 seconds.
+            guard lastTransaction == nil ||
+                  lastTransaction!.source != source || lastTransaction!.destination != destination ||
+                  request.isNewTransaction(previousTid: lastTransaction!.tid, timestamp: lastTransaction!.timestamp) else {
+                lastTransaction = (source: source, destination: destination, tid: request.tid, timestamp: Date())
+                break
+            }
+            lastTransaction = (source: source, destination: destination, tid: request.tid, timestamp: Date())
+            return DoozEpochStatus(tzData: request.mTzData, command: request.mCommand, io: request.mIO, unused: request.mUnused, epoch: request.mEpoch, correlation: request.mCorrelation, extra: request.mExtra)
         default:
             // Not possible.
             break